LayoutTests/accessibility/mac/supports-focus-setting-expected.txt

focusable link non focusable link focusable link non focusable button focusable button
This tests whether focus can be set is reported correctly based on whether the node can accept focus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS group.childAtIndex(0).isFocusable is true
PASS group.childAtIndex(1).isFocusable is false
PASS group.childAtIndex(2).isFocusable is true
PASS group.childAtIndex(3).isFocusable is false
PASS group.childAtIndex(4).isFocusable is true
PASS successfullyParsed is true

TEST COMPLETE
